requirements.txt
# Core Django Framework
django

# Production Server & Static Files
gunicorn
whitenoise

# Database URL Parsing & Connection
dj_database_url
psycopg2-binary

# Environment Variables
python-decouple

# Google AI (Gemini) & Vertex AI
google-generativeai >= 0.5.0 # Includes embedding generation
google-cloud-aiplatform >= 1.38.0 # Includes Matching Engine client
# Core Google Cloud dependencies (often needed)
google-api-core >= 2.11.0
google-auth >= 2.15.0
google-cloud-core >= 2.3.0
google-cloud-storage >= 2.5.0 # Often used by aiplatform for auth/project context
googleapis-common-protos >= 1.56.0
proto-plus >= 1.22.0
protobuf >= 3.19.5 # Ensure compatible protobuf

# Other common dependencies (might be pulled in automatically, but safe to list)
requests >= 2.28.0
urllib3 >= 1.26 # Often needed by requests/google libs
grpcio >= 1.47.0 # Needed by google-cloud libs
grpcio-status # Needed by google-cloud libs
cachetools # Often needed by google-auth
pyasn1-modules # Often needed by google-auth
rsa # Often needed by google-auth
six # Very common utility
python-dateutil # Common utility
sqlparse # Django dependency
asgiref # Django dependency
typing_extensions # Used by many modern libraries
numpy # Often needed by ML/vector libraries, even clients
redis>=4.0.0,<5.0.0  
celery>=5.0,<6.0   